+++ Assets/EditorScripts/NotNull/Editor/DisallowPlayModeIfNull.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 37a7f25782506e14f9558a0f6fb47f40
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
